Blogger v3.04 (User roles)

+++ VebAPI/Controllers/V1/PostsController.cs
@@ -1,9 +1,12 @@
 ﻿using Application.Dto;
 using Application.Interfaces;
+using Infrastructure.Identity;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.OData.Query;
 using Swashbuckle.AspNetCore.Annotations;
+using System.Security.Claims;
 using WebAPI.Filters;
 using WebAPI.Helpers;
 using WebAPI.Wrappers;
@@ -13,6 +16,7 @@ namespace WebAPI.Controllers.V1;
 //[ApiExplorerSettings(IgnoreApi = true)]
 [Route("api/[controller]")]
 [ApiVersion("1.0")]
+[Authorize]
 [ApiController]
 //[ApiExplorerSettings(GroupName = "v1")]
 public class PostsController : ControllerBase
@@ -23,15 +27,15 @@ public class PostsController : ControllerBase
         _postService = postService;
     }
 
-    [SwaggerOperation(Summary = "Retrieves sort fields")]
-    [HttpGet("[action]")]
-    public IActionResult GetSortFields()
-    {
-        return Ok(SortingHelper.GetSortFields().Select(x => x.Key));
-    }
+    //[SwaggerOperation(Summary = "Retrieves sort fields")]
+    //[HttpGet("[action]")]
+    //public IActionResult GetSortFields()
+    //{
+    //    return Ok(SortingHelper.GetSortFields().Select(x => x.Key));
+    //}
 
     [SwaggerOperation(Summary = "Retrieves paged posts")]
-    [HttpGet]
+    [HttpGet("Get")]
     public async Task<IActionResult> Get([FromQuery] PaginationFilter paginationFilter, [FromQuery] SortingFilter sortingFilter, [FromQuery] string filterBy = "")
     {
         var validPaginationFilter = new PaginationFilter(paginationFilter.PageNumber, paginationFilter.PageSize);
@@ -46,8 +50,10 @@ public class PostsController : ControllerBase
     }
 
     [SwaggerOperation(Summary ="Retrives all posts")]
-    [EnableQuery]
+    [Authorize(Roles = UserRoles.Admin)]
     [HttpGet("[action]")]
+    //[HttpGet("GetAll")]
+    [EnableQuery]
     public IQueryable<PostDto> GetAll()
     {
         return _postService.GetAllPosts();
@@ -66,26 +72,44 @@ public class PostsController : ControllerBase
         return Ok(new Response<PostDto>(post));
     }
 
+
     [SwaggerOperation(Summary = "Create a new post")]
+    [Authorize(Roles = UserRoles.User)]
     [HttpPost]
     public async Task<IActionResult> Create(CreatePostDto newPost)
     {
-        var post = await _postService.AddNewPostAsync(newPost);
+        var post = await _postService.AddNewPostAsync(newPost, User.FindFirstValue(ClaimTypes.NameIdentifier));
         return Created($"api/posts/{post.Id}",new Response<PostDto>(post));
     }
 
     [SwaggerOperation(Summary = "Update a exsisting post")]
+    [Authorize(Roles = UserRoles.User)]
     [HttpPut]
     public async Task<IActionResult> Update(UpdatePostDto updatePost)
     {
+        var userOwnsPost = await _postService.UserOwnsPostAsync(updatePost.Id, User.FindFirstValue(ClaimTypes.NameIdentifier));
+        if (!userOwnsPost)
+        {
+            return BadRequest(new Response<bool>() { Succeeded = false, Message = "You do not own this post" });
+        }
+
         await _postService.UpdatePostAsync(updatePost);
         return NoContent();
     }
 
     [SwaggerOperation(Summary = "Delete a specific post")]
-    [HttpDelete("Id")]
+    [Authorize(Roles = UserRoles.AdminOrUser)]
+    [HttpDelete("{id}")]
     public async Task<IActionResult> Delete(int id)
     {
+        var userOwnsPost = await _postService.UserOwnsPostAsync(id, User.FindFirstValue(ClaimTypes.NameIdentifier));
+        var isAdmin = User.FindFirstValue(ClaimTypes.Role).Contains(UserRoles.Admin);
+        
+        if (!isAdmin && !userOwnsPost)
+        {
+            return BadRequest(new Response<bool>() { Succeeded = false, Message = "You do not own this post" });
+        }
+
         await _postService.DeletePostAsync(id);
         return NoContent();
     }
